Engine: 	- removed 32bit support 	- removed UWP support 	- Windows target now requires Vista or newer 	- new rendering abstraction stub 	- replaced atomic wrapper with stdatomic

+++ Deps/PhysFS/physfs_platform_apple.m
@@ -159,7 +159,7 @@ void __PHYSFS_platformDetectAvailableCDs(PHYSFS_StringCallback cb, void *data)
 {
 #if !defined(PHYSFS_NO_CDROM_SUPPORT)
     const char *devPrefix = "/dev/";
-    const int prefixLen = strlen(devPrefix);
+    const int prefixLen = (int)strlen(devPrefix);
     mach_port_t masterPort = 0;
     struct statfs *mntbufp;
     int i, mounts;
